Update Configuration.h
Added config for a second extruder. Removed bed config (since a bed and second extruder aren't compatible).

Updated the RAMPS pinmap to EEF from EFB (necessary to have the second extruder act as a second extruder; the other option would have been to define the extruder as the bed pins).